🚅 perfs: optimize psbtUtil to other dir
