+++ src/psbt/psbtutils.d.ts
@@ -1,31 +1,7 @@
 /// <reference types="node" />
 import { PsbtInput } from 'bip174/src/lib/interfaces';
-/**
- * Converts a witness stack to a script witness.
- * @param witness The witness stack to convert.
- * @returns The script witness as a Buffer.
- */
-/**
- * Converts a witness stack to a script witness.
- * @param witness The witness stack to convert.
- * @returns The converted script witness.
- */
-export declare function witnessStackToScriptWitness(witness: Buffer[]): Buffer;
-/**
- * Finds the position of a public key in a script.
- * @param pubkey The public key to search for.
- * @param script The script to search in.
- * @returns The index of the public key in the script, or -1 if not found.
- * @throws {Error} If there is an unknown script error.
- */
-export declare function pubkeyPositionInScript(pubkey: Buffer, script: Buffer): number;
-/**
- * Checks if a public key is present in a script.
- * @param pubkey The public key to check.
- * @param script The script to search in.
- * @returns A boolean indicating whether the public key is present in the script.
- */
-export declare function pubkeyInScript(pubkey: Buffer, script: Buffer): boolean;
+import { HDSigner, HDSignerAsync, PsbtCache, PsbtOpts, Signer, SignerAsync } from './interfaces';
+import { Psbt } from '../psbt';
 /**
  * Checks if an input contains a signature for a specific action.
  * @param input - The input to check.
@@ -45,4 +21,9 @@ type SignatureDecodeFunc = (buffer: Buffer) => {
  * @returns True if the action is allowed, false otherwise.
  */
 export declare function signatureBlocksAction(signature: Buffer, signatureDecodeFn: SignatureDecodeFunc, action: string): boolean;
+export declare function check32Bit(num: number): void;
+export declare function checkFees(psbt: Psbt, cache: PsbtCache, opts: PsbtOpts): void;
+export declare function getSignersFromHD(inputIndex: number, inputs: PsbtInput[], hdKeyPair: HDSigner | HDSignerAsync): Array<Signer | SignerAsync>;
+export declare function range(n: number): number[];
+export declare function isPubkeyLike(buf: Buffer): boolean;
 export {};
