+++ ts_src/psbt.ts
@@ -40,7 +40,6 @@ import {
   checkTaprootInputFields,
   checkTaprootOutputFields,
 } from './psbt/bip371';
-import { witnessStackToScriptWitness } from './psbt/psbtutils';
 import { PsbtTransaction, transactionFromBuffer } from './psbt/transaction';
 import {
   addNonWitnessTxCache,
@@ -59,7 +58,7 @@ import {
   inputFinalizeGetAmts,
   pubkeyInInput,
 } from './psbt/input';
-import { checkPartialSigSighashes } from './psbt/global/sign';
+import { checkPartialSigSighashes, trimTaprootSig } from './psbt/global/sign';
 import {
   checkInvalidP2WSH,
   checkScriptForPubkey,
@@ -67,7 +66,9 @@ import {
   getMeaningfulScript,
   getScriptFromInput,
   getScriptFromUtxo,
-  scriptWitnessToWitnessStack,
+  redeemFromFinalScriptSig,
+  redeemFromFinalWitnessScript,
+  witnessStackToScriptWitness,
 } from './psbt/input/script';
 import { pubkeyInOutput } from './psbt/output';
 import {
@@ -76,6 +77,12 @@ import {
   getHashForSig,
   getTaprootHashesForSig,
 } from './psbt/global/hash';
+import {
+  check32Bit,
+  checkFees,
+  getSignersFromHD,
+  range,
+} from './psbt/psbtutils';
 
 /**
  * These are the default arguments for a Psbt instance.
@@ -1137,109 +1144,4 @@ export class Psbt {
     this.data.clearFinalizedInput(inputIndex);
     return this;
   }
-}
-
-function check32Bit(num: number): void {
-  if (
-    typeof num !== 'number' ||
-    num !== Math.floor(num) ||
-    num > 0xffffffff ||
-    num < 0
-  ) {
-    throw new Error('Invalid 32 bit integer');
-  }
-}
-
-function checkFees(psbt: Psbt, cache: PsbtCache, opts: PsbtOpts): void {
-  const feeRate = cache.__FEE_RATE || psbt.getFeeRate();
-  const vsize = cache.__EXTRACTED_TX!.virtualSize();
-  const satoshis = feeRate * vsize;
-  if (feeRate >= opts.maximumFeeRate) {
-    throw new Error(
-      `Warning: You are paying around ${(satoshis / 1e8).toFixed(8)} in ` +
-        `fees, which is ${feeRate} satoshi per byte for a transaction ` +
-        `with a VSize of ${vsize} bytes (segwit counted as 0.25 byte per ` +
-        `byte). Use setMaximumFeeRate method to raise your threshold, or ` +
-        `pass true to the first arg of extractTransaction.`,
-    );
-  }
-}
-
-function trimTaprootSig(signature: Buffer): Buffer {
-  return signature.length === 64 ? signature : signature.subarray(0, 64);
-}
-
-function getSignersFromHD(
-  inputIndex: number,
-  inputs: PsbtInput[],
-  hdKeyPair: HDSigner | HDSignerAsync,
-): Array<Signer | SignerAsync> {
-  const input = checkForInput(inputs, inputIndex);
-  if (!input.bip32Derivation || input.bip32Derivation.length === 0) {
-    throw new Error('Need bip32Derivation to sign with HD');
-  }
-  const myDerivations = input.bip32Derivation
-    .map(bipDv => {
-      if (bipDv.masterFingerprint.equals(hdKeyPair.fingerprint)) {
-        return bipDv;
-      } else {
-        return;
-      }
-    })
-    .filter(v => !!v);
-  if (myDerivations.length === 0) {
-    throw new Error(
-      'Need one bip32Derivation masterFingerprint to match the HDSigner fingerprint',
-    );
-  }
-  const signers: Array<Signer | SignerAsync> = myDerivations.map(bipDv => {
-    const node = hdKeyPair.derivePath(bipDv!.path);
-    if (!bipDv!.pubkey.equals(node.publicKey)) {
-      throw new Error('pubkey did not match bip32Derivation');
-    }
-    return node;
-  });
-  return signers;
-}
-
-function redeemFromFinalScriptSig(
-  finalScript: Buffer | undefined,
-): Buffer | undefined {
-  if (!finalScript) return;
-  const decomp = bscript.decompile(finalScript);
-  if (!decomp) return;
-  const lastItem = decomp[decomp.length - 1];
-  if (
-    !Buffer.isBuffer(lastItem) ||
-    isPubkeyLike(lastItem) ||
-    isSigLike(lastItem)
-  )
-    return;
-  const sDecomp = bscript.decompile(lastItem);
-  if (!sDecomp) return;
-  return lastItem;
-}
-
-function redeemFromFinalWitnessScript(
-  finalScript: Buffer | undefined,
-): Buffer | undefined {
-  if (!finalScript) return;
-  const decomp = scriptWitnessToWitnessStack(finalScript);
-  const lastItem = decomp[decomp.length - 1];
-  if (isPubkeyLike(lastItem)) return;
-  const sDecomp = bscript.decompile(lastItem);
-  if (!sDecomp) return;
-  return lastItem;
-}
-
-function isPubkeyLike(buf: Buffer): boolean {
-  return buf.length === 33 && bscript.isCanonicalPubKey(buf);
-}
-
-function isSigLike(buf: Buffer): boolean {
-  return bscript.isCanonicalScriptSignature(buf);
-}
-
-function range(n: number): number[] {
-  return [...Array(n).keys()];
 }
