+++ ts_src/psbt/input/script.ts
@@ -2,10 +2,14 @@ import { PsbtInput } from "bip174/src/lib/interfaces";
 import { varuint } from "../../bufferutils";
 import * as  payments from "../../payments";
 import { GetScriptReturn, PsbtCache, ScriptType } from "../interfaces";
-import { pubkeyInScript } from "../psbtutils";
 import { nonWitnessUtxoTxFromCache } from "../global/cache";
+import { hash160 } from "../../crypto";
+import * as bscript from '../../script';
+import { isPubkeyLike } from "../psbtutils";
+import { isSigLike } from "../global/sign";
 
 const {isP2MS, isP2PK, isP2PKH, isP2SHScript, isP2WPKH, isP2WSHScript} = payments;
+
 export function getMeaningfulScript(
   script: Buffer,
   index: number,
@@ -179,6 +183,104 @@ export function getScriptFromInput(
   return res;
 }
 
+/**
+ * Converts a witness stack to a script witness.
+ * @param witness The witness stack to convert.
+ * @returns The converted script witness.
+ */
+export function witnessStackToScriptWitness(witness: Buffer[]): Buffer {
+  let buffer = Buffer.allocUnsafe(0);
+
+  function writeSlice(slice: Buffer): void {
+    buffer = Buffer.concat([buffer, Buffer.from(slice)]);
+  }
+
+  function writeVarInt(i: number): void {
+    const currentLen = buffer.length;
+    const varintLen = varuint.encodingLength(i);
+
+    buffer = Buffer.concat([buffer, Buffer.allocUnsafe(varintLen)]);
+    varuint.encode(i, buffer, currentLen);
+  }
+
+  function writeVarSlice(slice: Buffer): void {
+    writeVarInt(slice.length);
+    writeSlice(slice);
+  }
+
+  function writeVector(vector: Buffer[]): void {
+    writeVarInt(vector.length);
+    vector.forEach(writeVarSlice);
+  }
+
+  writeVector(witness);
+
+  return buffer;
+}
+
+/**
+ * Finds the position of a public key in a script.
+ * @param pubkey The public key to search for.
+ * @param script The script to search in.
+ * @returns The index of the public key in the script, or -1 if not found.
+ * @throws {Error} If there is an unknown script error.
+ */
+export function pubkeyPositionInScript(pubkey: Buffer, script: Buffer): number {
+  const pubkeyHash = hash160(pubkey);
+  const pubkeyXOnly = pubkey.slice(1, 33); // slice before calling?
+
+  const decompiled = bscript.decompile(script);
+  if (decompiled === null) throw new Error('Unknown script error');
+
+  return decompiled.findIndex(element => {
+    if (typeof element === 'number') return false;
+    return (
+      element.equals(pubkey) ||
+      element.equals(pubkeyHash) ||
+      element.equals(pubkeyXOnly)
+    );
+  });
+}
+
+/**
+ * Checks if a public key is present in a script.
+ * @param pubkey The public key to check.
+ * @param script The script to search in.
+ * @returns A boolean indicating whether the public key is present in the script.
+ */
+export function pubkeyInScript(pubkey: Buffer, script: Buffer): boolean {
+  return pubkeyPositionInScript(pubkey, script) !== -1;
+}
+
+export function redeemFromFinalScriptSig(
+  finalScript: Buffer | undefined,
+): Buffer | undefined {
+  if (!finalScript) return;
+  const decomp = bscript.decompile(finalScript);
+  if (!decomp) return;
+  const lastItem = decomp[decomp.length - 1];
+  if (
+    !Buffer.isBuffer(lastItem) ||
+    isPubkeyLike(lastItem) ||
+    isSigLike(lastItem)
+  )
+    return;
+  const sDecomp = bscript.decompile(lastItem);
+  if (!sDecomp) return;
+  return lastItem;
+}
+
+export function redeemFromFinalWitnessScript(
+  finalScript: Buffer | undefined,
+): Buffer | undefined {
+  if (!finalScript) return;
+  const decomp = scriptWitnessToWitnessStack(finalScript);
+  const lastItem = decomp[decomp.length - 1];
+  if (isPubkeyLike(lastItem)) return;
+  const sDecomp = bscript.decompile(lastItem);
+  if (!sDecomp) return;
+  return lastItem;
+}
 
 function scriptCheckerFactory(
   payment: any,
